requirements.txt
ecologits[mistralai]
langchain-mistralai
flask
python-dotenv
faiss-cpu
langchain-community
